Flatten home structure
