--- src/lib/web/modules/home/breweries/breweries.module.ts
@@ -1,11 +1,0 @@
-import { Module } from '@nestjs/common'
-
-import { BreweryModule as CoreBreweryModule } from '@app/core/brewery'
-
-import { HomeBreweriesController } from './breweries.controller'
-
-@Module({
-  imports: [CoreBreweryModule],
-  controllers: [HomeBreweriesController],
-})
-export class HomeBreweriesModule {}
