+++ src/lib/web/modules/home/home.controller.ts
@@ -1,10 +1,18 @@
 import { Controller, Get, Render, Res } from '@nestjs/common'
 
-@Controller('/home')
+import { BreweryService } from '@app/core/brewery'
+
+@Controller('/')
 class HomeController {
+  constructor(private readonly breweryService: BreweryService) {}
+
   @Get()
-  root(@Res() res) {
-    res.render('home/index', {
+  async index(@Res() res) {
+    const breweries = await this.breweryService.findAll()
+
+    res.render('home/breweries/index', {
+      breweries,
+      breweriesJson: JSON.stringify(breweries),
       layout: 'layouts/home_layout',
     })
   }
