+++ src/lib/web/modules/home/home.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common'
 import { HomeController } from './home.controller'
-import { HomeBreweriesModule } from './breweries/breweries.module'
+import { BreweryModule as CoreBreweryModule } from '@app/core/brewery'
 
 @Module({
-  imports: [HomeBreweriesModule],
+  imports: [CoreBreweryModule],
   controllers: [HomeController],
 })
 export class HomeModule {}
